requirements.txt
# Core Framework
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

# LLM
openai>=1.12.0

# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# File Processing
pytesseract>=0.3.10
PyPDF2>=3.0.0
Pillow>=10.0.0

# Audio Processing
openai-whisper>=20231117

# YouTube
youtube-transcript-api>=0.6.1

# Utilities
python-dotenv>=1.0.0